Added ability to move players in the list
Rows can be moved up/down in the list. Saving will save it in the order
the table is in.
Also made apiReq accept a callback as the second argument instead of in
the query object. Alleviates the need to call apiReq() after writing the
callback, making it more in line with other such functions
(xmlHTTPrequest, etc)